refactor(assets): extract create/read core into lib/asset-core.ts
Pure relocation of the versioned-asset creation + read paths out of
asset-service.ts into a new base module asset-core.ts: createAsset, getAsset,
assetExists, resolveFile/resolveFileFromManifest, getAssetSummary, listAssets,
the AssetCreateInput/AssetFileRef/AssetSummary interfaces, the allocateAssetDir/
toSummary internals, and the shared leaf helpers extOf + nowIso (the latter two
exported so the mutation/upsert clusters can import them).

asset-core is the cycle-free base: it imports no sibling. asset-service.ts
re-exports the create/read public surface so the 21 consumers + tests keep
importing from `…/lib/asset-service` unchanged, and continues to house the
mutation + upsert code (which now imports getAsset/listAssets/createAsset/extOf/
nowIso from asset-core, never the barrel). Dropped the now-unused extname/
generateAssetId/AssetSource imports from asset-service. Function bodies
byte-for-byte unchanged.

Verified: typecheck/eslint/assets suite 242-0/full suite 5124-0/binary build
(stamps reverted) + isolated boot smoke (assets plugin activates; the search
reindex runs listAssets clean).
